+++ packages/server/lib/helpers/validation.ts
@@ -88,7 +88,7 @@ export const integrationCredentialsSchema = z.discriminatedUnion(
 
 export const sharedCredentialsSchema = z
     .object({
-        name: z.string().min(1).max(255),
+        name: providerNameSchema,
         client_id: z.string().min(1).max(255),
         client_secret: z.string().min(1),
         scopes: z.union([z.string().regex(/^[0-9a-zA-Z:/_.-]+(,[0-9a-zA-Z:/_.-]+)*$/), z.string().max(0)]).optional()
