+++ packages/server/lib/routes.internal.ts
@@ -2,10 +2,10 @@ import bodyParser from 'body-parser';
 import express from 'express';
 
 import { postRollout } from './controllers/fleet/postRollout.js';
-import { getSharedCredentialsProviders } from './controllers/v1/integrations/sharedCredentials/getListSharedCredentials.js';
-import { getSharedCredentialsProvider } from './controllers/v1/integrations/sharedCredentials/getSharedCredential.js';
-import { patchSharedCredentialsProvider } from './controllers/v1/integrations/sharedCredentials/patchSharedCredentialsProvider.js';
-import { postSharedCredentialsProvider } from './controllers/v1/integrations/sharedCredentials/postSharedCredentialsProvider.js';
+import { getSharedCredentialsProviders } from './controllers/sharedCredentials/getListSharedCredentials.js';
+import { getSharedCredentialsProvider } from './controllers/sharedCredentials/getSharedCredential.js';
+import { patchSharedCredentialsProvider } from './controllers/sharedCredentials/patchSharedCredential.js';
+import { postSharedCredentialsProvider } from './controllers/sharedCredentials/postSharedCredential.js';
 import authMiddleware from './middleware/access.middleware.js';
 import { rateLimiterMiddleware } from './middleware/ratelimit.middleware.js';
 
@@ -21,7 +21,7 @@ internalApi.use(bodyParser.raw({ limit: bodyLimit }));
 const interalApiAuth: RequestHandler[] = [rateLimiterMiddleware, authMiddleware.internal.bind(authMiddleware)];
 internalApi.route('/fleet/:fleetId/rollout').post(interalApiAuth, postRollout);
 
-internalApi.route('/shared-credentials/providers').get(interalApiAuth, getSharedCredentialsProviders);
-internalApi.route('/shared-credentials/providers/:name').get(interalApiAuth, getSharedCredentialsProvider);
-internalApi.route('/shared-credentials/providers').post(interalApiAuth, postSharedCredentialsProvider);
-internalApi.route('/shared-credentials/providers/:id').patch(interalApiAuth, patchSharedCredentialsProvider);
+internalApi.route('/shared-credentials').get(interalApiAuth, getSharedCredentialsProviders);
+internalApi.route('/shared-credentials/:name').get(interalApiAuth, getSharedCredentialsProvider);
+internalApi.route('/shared-credentials').post(interalApiAuth, postSharedCredentialsProvider);
+internalApi.route('/shared-credentials/:name').patch(interalApiAuth, patchSharedCredentialsProvider);
