+++ packages/types/lib/sharedCredentials/api.ts
@@ -2,7 +2,7 @@ import type { ApiError, Endpoint } from '../api.js';
 
 export type GetSharedCredentialsProviders = Endpoint<{
     Method: 'GET';
-    Path: '/internal/shared-credentials/providers';
+    Path: '/internal/shared-credentials';
     Success: {
         success: boolean;
         data: SharedCredentialsOutput[];
@@ -11,7 +11,7 @@ export type GetSharedCredentialsProviders = Endpoint<{
 
 export type GetSharedCredentialsProvider = Endpoint<{
     Method: 'GET';
-    Path: '/internal/shared-credentials/providers/:name';
+    Path: '/internal/shared-credentials/:name';
     Params: { name: string };
     Success: {
         success: boolean;
@@ -21,27 +21,27 @@ export type GetSharedCredentialsProvider = Endpoint<{
 
 export type PostSharedCredentialsProvider = Endpoint<{
     Method: 'POST';
-    Path: '/internal/shared-credentials/providers';
+    Path: '/internal/shared-credentials';
     Body: SharedCredentialsBodyInput;
     Success: {
         success: boolean;
     };
-    Error: ApiError<'invalid_body' | 'shared_credentials_already_exists'>;
+    Error: ApiError<'invalid_body' | 'shared_credentials_already_exists' | 'invalid_provider'>;
 }>;
 
 export type PatchSharedCredentialsProvider = Endpoint<{
     Method: 'PATCH';
-    Path: '/internal/shared-credentials/providers/:id';
-    Params: { id: string };
-    Body: SharedCredentialsBodyInput & { name: string };
+    Path: '/internal/shared-credentials/:name';
+    Params: { name: string };
+    Body: SharedCredentialsBodyInput;
     Success: {
         success: boolean;
     };
-    Error: ApiError<'invalid_body' | 'shared_credentials_provider_not_found' | 'shared_credentials_already_exists'>;
+    Error: ApiError<'invalid_body' | 'shared_credentials_provider_not_found' | 'shared_credentials_already_exists' | 'invalid_provider'>;
 }>;
 
 export interface SharedCredentialsBodyInput {
-    name?: string | undefined;
+    name: string;
     client_id: string;
     client_secret: string;
     scopes?: string | undefined;
